move keyvalue eq and hash to common
